+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Allow to specify directory where to output repacked phar
 * Do not ask confirmation when explicitly using `castor init` command
+* Allow to pass a callback when using the `run_php` function (similar to the `run` function)
 
 ### Fixes
 
@@ -16,6 +17,7 @@
 * Provide PHP 8.4 in the static binaries
 * Update all PHP vendor
 * Upgrade SPC version to v2.7.5
+* Use a `PhpRunner` service instead of static functions to run PHP files
 
 ## 1.0.0 (2025-10-10)